+++ src/main/java/com/blog/som/domain/member/security/service/AuthService.java
@@ -8,8 +8,7 @@ import com.blog.som.domain.member.dto.TokenResponse;
 import com.blog.som.domain.member.entity.MemberEntity;
 import com.blog.som.domain.member.repository.MemberRepository;
 import com.blog.som.domain.member.security.token.JwtTokenService;
-import com.blog.som.domain.member.type.Role;
-import com.blog.som.global.components.password.PasswordUtils;
+import com.blog.som.global.util.PasswordUtils;
 import com.blog.som.global.exception.ErrorCode;
 import com.blog.som.global.exception.custom.MemberException;
 import com.blog.som.global.redis.token.TokenRepository;
